+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 ### Added
-- New argument to `search` to sort pages alphabetically
+- New argument to `search.pages()` to sort pages alphabetically
+
+### Removed
+- Arguments `path` and `recursive` in `search.pages()`
 
 ### Fixed
 - `url` filter with relative urls